requirements.txt
decord
fairscale==0.4.4
iopath
omegaconf
opencv-python-headless
pandas
scikit-image
sentencepiece
timm==0.4.12
tqdm
transformers==4.33.2
webdataset
torchaudio
moviepy==1.0.3
nltk
open3d
h5py
